Move multi-select state to a mixin
Not something we’ve utilized in the app before, but this pattern is in toolkit.

While we should only carefully use mixins, in this case this seems to be a very defined pattern.
